+++ cpp05/ex01/Form.cpp
@@ -6,7 +6,7 @@
 /*   By: dximenez <dximenez@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 18:31:23 by dximenez          #+#    #+#             */
-/*   Updated: 2024/08/08 21:41:45 by dximenez         ###   ########.fr       */
+/*   Updated: 2024/08/08 21:48:09 by dximenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,12 +52,12 @@ bool	Form::getSigned() const
 	return (_signed);
 }
 
-uint8_t	Form::getGradeToSign() const
+int	Form::getGradeToSign() const
 {
 	return (_gradeToSign);
 }
 
-uint8_t	Form::getGradeToExecute() const
+int	Form::getGradeToExecute() const
 {
 	return (_gradeToExecute);
 }
